# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - BearerAuth: []
...
paths:
...
  /residents/import/template:
    get:
      operationId: downloadResidentImportTemplate
      tags: [resident-import]
      summary: Download a resident import template
      security:
        - BearerAuth: []
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, xlsx]
      responses:
        "200":
          description: One-client-per-row import template
          content:
            text/csv: { schema: { type: string } }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: { schema: { type: string, format: binary } }
        "403": { description: Authorized staff access required }
  /residents/import/preview:
    post:
      operationId: previewResidentImport
      tags: [resident-import]
      summary: Parse and validate a resident import without saving residents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [filename, contentBase64]
              properties:
                filename: { type: string, minLength: 1, maxLength: 255 }
                contentBase64: { type: string, minLength: 1, maxLength: 1400000 }
      responses:
        "201":
          description: Row-by-row validation preview and persisted import batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentImportPreview"
        "400": { description: Invalid or unsupported upload }
        "403": { description: Authorized staff access required }
  /residents/import/{batchId}/confirm:
    post:
      operationId: confirmResidentImport
      tags: [resident-import]
      summary: Explicitly confirm valid resident import rows
      security:
        - BearerAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: number
            multipleOf: 1
            minimum: 1
            maximum: 2147483647
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [approvedRowNumbers]
              properties:
                approvedRowNumbers:
                  type: array
                  minItems: 1
                  items: { type: number, multipleOf: 1, minimum: 2 }
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentImportResult"
        "400": { description: No valid rows were approved }
        "403": { description: Authorized staff access required }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-session
...
  schemas:
...
    ResidentImportPreview:
      type: object
      required: [batchId, sourceFilename, identityRule, columns, rows, summary]
      properties:
        batchId: { type: number, multipleOf: 1 }
        sourceFilename: { type: string }
        identityRule: { type: string }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: object
            required: [rowNumber, sourceData, normalizedData, errors, valid]
            properties:
              rowNumber: { type: number, multipleOf: 1 }
              sourceData: { type: object, additionalProperties: { type: string } }
              normalizedData: { type: object, additionalProperties: true }
              errors: { type: array, items: { type: string } }
              valid: { type: boolean }
        summary:
          type: object
          required: [total, valid, failed]
          properties:
            total: { type: number, multipleOf: 1 }
            valid: { type: number, multipleOf: 1 }
            failed: { type: number, multipleOf: 1 }
    ResidentImportResult:
      type: object
      required: [batchId, status, importedResidentIds, imported, skipped]
      properties:
        batchId: { type: number, multipleOf: 1 }
        status: { type: string }
        importedResidentIds:
          type: array
          items: { type: number, multipleOf: 1 }
        imported: { type: number, multipleOf: 1 }
        skipped: { type: number, multipleOf: 1 }
...